requirements.txt
python-multipart
fastapi
pydantic
uvicorn

openai
google-genai

python-dotenv